feat: add button to toggle fit width and height
This also creates theme files for both OsdToolbar and ViewerHeader components, så we can set a background color bases on theme when a button is clicked. There will be som changes to this after merging with main and the new OSD is available.
Components still need formatting and possible some refactoring.