fix: swipe left when on last page causes incorrect zoom level when page is fitted to height
Solved by setting minZoom after changing viewer mode when updating fitToHeight
When updating fitToWidth, the zoom level also seems to be incorrect. However, it does not cause the same issue as when fitted to height. Just for code consistency, I'm also setting minZoom in this case, although it might not be needed.