refactor: goToCanvasGroupStrategy
remove unused code from vertical strategy. Vertical Strategy don't need to consider startOnTopOnCanvasGroupChange parameter. Because of this quit a bit of code could be removed.
calculation of next canvas group index when going to next or previous canvas had a lot of duplicate code with only Direction and ScrollDirection as difference. Extracted a lot of that code into a new function with directions as inputs.